moved the making of db tables to its own file
